requirements.txt
python-pptx==1.0.2
pillow==11.3.0
python-docx==1.2.0
PyYAML==6.0.2
requests==2.32.4
numpy==2.3.1
tqdm==4.67.1
pandas==2.3.1
pytictoc==1.5.3
opencv-python==4.10.0.84
ollama==0.5.1
psutil>=5.9.0
reportlab>=4.0.0
PyMuPDF>=1.23.0
rich>=13.0.0
jsonschema>=4.0.0

